% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLines2pdf.R
\name{addLines2pdf}
\alias{addLines2pdf}
\title{Add line numbers to a PDF}
\usage{
addLines2pdf(file)
}
\arguments{
\item{file}{the PDF file to add line numbmers to}
}
\value{
silently returns the path to a new file saved on the local computer
which is a copy of the file passed as argument \code{file} but with line
numbers
}
\description{
Add line numbers to a PDF
}
